Fix GCC error comparing std::string_views at compile time
This commit fixes a compile-time comparison.
GCC 10 refuses to compare std::string_views at compile-time with ==.
It declares that std::string_view is not a structural type.
This appears to be true according to the standard for C++17 but not 20.
Perhaps this has yet to updated as of GCC 10 for C++20.
Luckily, the built-in std::string_view compare function is constexpr.
So the workaround is straightforward.